Add Purchase model
Purchase will hold information on each purchase made in the system.
Besides the obvious references to the Purchaser and the InventoryItem,
it also has a reference to the Import file so that the gross value can
be displayed at the end and it also stores the price used for that
particular purchase.

This allows prices to change over time without "modifying history".  It
also allows for special discounts to be applyed to a given purchase,
even though the item price may be different.